requirements.txt
colorama
python-dotenv
lark
PyPDF2